requirements.txt
absl-py==2.1.0
accelerate==1.2.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asttokens==3.0.0
async-timeout==5.0.1
asyncer==0.0.8
attrs==24.3.0
augly==1.0.0
av==14.0.1
blinker==1.9.0
calflops==0.3.2
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.7
dearpygui==2.0.0
debugpy==1.8.11
decorator==5.1.1
diffusers==0.32.1
distlib==0.3.9
docstring_parser==0.16
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.13.1
filetype==1.2.0
Flask==3.0.3
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.2.0
ftfy==6.3.1
future==1.0.0
GPUtil==1.4.0
gradio==5.12.0
gradio_client==1.5.4
grpcio==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
identify==2.6.5
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jaxtyping==0.2.36
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiui==0.2.14
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
lazy_loader==0.4
lightning-utilities==0.11.9
llvmlite==0.43.0
loguru==0.7.3
lpips==0.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.25.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.3
objaverse==0.1.7
objprint==0.3.0
omegaconf==2.3.0
onnxruntime-gpu==1.20.1
open3d==0.19.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.14
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
plotly==5.24.1
plyfile==1.1
pooch==1.8.2
portalocker==3.1.1
pre_commit==4.0.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyav==14.0.1
pycocotools==2.0.8
pycryptodome==3.20.0
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
pygltflib==1.16.3
Pygments==2.19.1
PyMatting==1.1.13
pyparsing==3.2.1
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-multipart==0.0.20
pytorch-msssim==1.0.0
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
rembg==2.0.61
requests==2.32.3
retrying==1.3.4
rich==13.9.4
roma==1.5.1
rpds-py==0.22.3
ruff==0.9.1
safehttpx==0.1.6
safetensors==0.5.2
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.1
semantic-version==2.10.0
shellingham==1.5.4
shtab==1.7.1
six==1.17.0



torchmetrics==1.6.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.0
trimesh==4.5.3
triton==3.0.0
typeguard==4.4.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tyro==0.9.9
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
varname==0.13.5
virtualenv==20.29.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.17.2
xatlas==0.0.9
xformers==0.0.27.post2
yarl==1.18.3
zipp==3.21.0